Remove success alerts and logic in useAlerts
