+++ packages/checkout-storefront/src/hooks/useAlerts/types.ts
@@ -1,10 +1,6 @@
 import { ErrorCode } from "@/checkout-storefront/lib/globalTypes";
 
-export type AlertType = "success" | "error";
-
-export interface AlertSuccessData {
-  scope: CheckoutScope;
-}
+export type AlertType = "error";
 
 export interface AlertErrorData {
   scope: CheckoutScope;
