+++ packages/checkout-storefront/src/hooks/useAlerts/useAlerts.tsx
@@ -1,6 +1,6 @@
 import { useErrorMessages } from "@/checkout-storefront/hooks/useErrorMessages";
 import { MessageKey, useFormattedMessages } from "@/checkout-storefront/hooks/useFormattedMessages";
-import { Alert, AlertType, AlertErrorData, AlertSuccessData, CheckoutScope } from "./types";
+import { Alert, AlertType, AlertErrorData, CheckoutScope } from "./types";
 import { toast } from "react-toastify";
 import { camelCase } from "lodash-es";
 import { ApiErrors, useGetParsedApiErrors } from "@/checkout-storefront/hooks/useErrors";
@@ -8,7 +8,6 @@ import { ErrorCode } from "@/checkout-storefront/lib/globalTypes";
 import { warnAboutMissingTranslation } from "../useFormattedMessages/utils";
 
 export interface ScopedAlertsProps {
-  showSuccess: () => void;
   showErrors: (errors: AlertErrorData[]) => void;
   showCustomErrors: (errors: ErrorCode[]) => void;
 }
@@ -16,7 +15,6 @@ export interface ScopedAlertsProps {
 function useAlerts(scope: CheckoutScope): ScopedAlertsProps;
 
 function useAlerts(): {
-  showSuccess: (scope: CheckoutScope) => void;
   showErrors: (errors: AlertErrorData[], scope: CheckoutScope) => void;
   showCustomErrors: (errors: ErrorCode[], scope: CheckoutScope) => void;
 };
@@ -40,38 +38,20 @@ function useAlerts(globalScope?: any): any {
     }
   };
 
-  const getParsedAlert = (data: AlertSuccessData | AlertErrorData, type: AlertType): Alert => {
-    if (type === "error") {
-      const { scope, field, code } = data as AlertErrorData;
-
-      return {
-        id: camelCase(`${scope}-${field}-${code}`),
-        message: getErrorMessage({ scope, code, field }),
-        type,
-      };
-    }
-
-    const { scope } = data;
-
-    const id = `${scope}Success`;
+  const getParsedAlert = (data: AlertErrorData, type: AlertType): Alert => {
+    const { scope, field, code } = data as AlertErrorData;
 
     return {
+      id: camelCase(`${scope}-${field}-${code}`),
+      message: getErrorMessage({ scope, code, field }),
       type,
-      id,
-      message: formatMessage(id as MessageKey),
     };
   };
 
-  const showAlert = (
-    { scope, ...dataRest }: AlertSuccessData | AlertErrorData,
-    { type }: { type: AlertType }
-  ) => {
+  const showAlert = ({ scope, ...dataRest }: AlertErrorData, { type }: { type: AlertType }) => {
     const { message, ...options } = getParsedAlert({ ...dataRest, scope }, type);
     toast(message, options);
   };
-
-  const showSuccess = (scope: CheckoutScope = globalScope) =>
-    showAlert({ scope }, { type: "success" });
 
   const showErrors = (errors: ApiErrors<any>, scope: CheckoutScope = globalScope) =>
     getParsedApiErrors(errors).forEach((error) =>
@@ -81,7 +61,7 @@ function useAlerts(globalScope?: any): any {
   const showCustomErrors = (errors: ErrorCode[], scope: CheckoutScope = globalScope) =>
     errors.forEach((code: ErrorCode) => showAlert({ scope, code, field: "" }, { type: "error" }));
 
-  return { showSuccess, showErrors, showCustomErrors };
+  return { showErrors, showCustomErrors };
 }
 
 export { useAlerts };
